PLATUI-1336: Refactored Utils following code review
