+++ src/main/scala/uk/gov/hmrc/supportfrontend/views/UtilsSupport.scala
@@ -22,7 +22,7 @@ import uk.gov.hmrc.hmrcfrontend.views.Implicits.RichHtml
 
 import java.net.URLEncoder
 
-trait Utils {
+trait UtilsSupport {
 
   /**
     * Creates a space-separated list of CSS classes to be included in a template.
@@ -48,16 +48,6 @@ trait Utils {
     htmlArgs.padLeft(if (attributes.nonEmpty) padCount else 0)
   }
 
-  def isNonEmptyOptionString(value: Option[String]): Boolean = value match {
-    case Some(NonEmptyString(_)) => true
-    case _                       => false
-  }
-
-  def calculateAssetPath(path: Option[String], file: String): String =
-    path
-      .map(p => s"$p/$file")
-      .getOrElse(uk.gov.hmrc.govukfrontend.controllers.routes.Assets.at(file).url)
-
   object NonEmptyString {
     def unapply(s: String): Option[String] =
       if (s != null && s.nonEmpty) Some(s) else None
@@ -65,6 +55,5 @@ trait Utils {
 
   def urlEncode(s: String): String =
     URLEncoder.encode(s, "UTF-8").replace("+", "%20")
-}
 
-object Utils extends Utils
+}
