+++ src/main/twirl/uk/gov/hmrc/govukfrontend/views/components/GovukTemplate.scala.html
@@ -16,7 +16,6 @@
 
 @import uk.gov.hmrc.govukfrontend.controllers.routes
 @import uk.gov.hmrc.govukfrontend.views.html.components._
-@import uk.gov.hmrc.supportfrontend.views.Utils._
 
 @this(govukHeader: GovukHeader, govukFooter: GovukFooter, govukSkipLink: GovukSkipLink)
